pyproject.toml
[project]
name = "hts-classifier"
version = "0.1.0"
description = "HTS tariff code classifier using AI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "google-genai>=1.70.0",
    "chromadb>=0.5.0",
    "pydantic-settings>=2.5.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.26.0",
    "rank-bm25>=0.2.2",
    "httpx>=0.27.0",
    "tqdm>=4.66.0",
    "loguru>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
# onnxruntime 1.20+ dropped Intel Mac wheels; pin to last version with x86_64 support
override-dependencies = ["onnxruntime<1.20"]

[dependency-groups]
dev = [
    "ruff>=0.15.9",
    "ty>=0.0.28",
]
extend-select = [
    "I",   # isort
]